+++ pkg/cmd/attach/cmd.go
@@ -4,6 +4,7 @@ package attach
 import (
 	"github.com/stolostron/cm-cli/pkg/cmd/attach/cluster"
 	"github.com/stolostron/cm-cli/pkg/cmd/attach/clusterclaim"
+	"github.com/stolostron/cm-cli/pkg/cmd/attach/hostedcluster"
 	genericclioptionscm "github.com/stolostron/cm-cli/pkg/genericclioptions"
 
 	"github.com/spf13/cobra"
@@ -19,6 +20,7 @@ func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOSt
 
 	cmd.AddCommand(cluster.NewCmd(cmFlags, streams))
 	cmd.AddCommand(clusterclaim.NewCmd(cmFlags, streams))
+	cmd.AddCommand(hostedcluster.NewCmd(cmFlags, streams))
 
 	return cmd
 }