+++ pkg/cmd/detach/cluster/cmd.go
@@ -39,8 +39,8 @@ func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOSt
 
 	cluster := &cobra.Command{
 		Use:          "cluster",
-		Aliases:      []string{"clusters", "clusterclaim", "clusterclaims", "cc", "ccs"},
-		Short:        "detach a cluster",
+		Aliases:      []string{"clusters", "clusterclaim", "clusterclaims", "cc", "ccs", "hostedcluster", "hostedclusters", "hc", "hcs"},
+		Short:        "detach a cluster, clusterclaim or hostedcluster",
 		Example:      fmt.Sprintf(example, helpers.GetExampleHeader()),
 		SilenceUsage: true,
 		PreRunE: func(c *cobra.Command, args []string) error {