+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Additions
 
 - [cm attach should handle kubeconfig blobs #173](https://github.com/stolostron/cm-cli/issues/173)
-
+- [Create a cmd cm attach/detach hostedcluster #182](https://github.com/stolostron/cm-cli/issues/182)
 ## Breaking changes
 
 ## Bug fixes